# Conflicts:
#	src/main.kt

+++ src/Models/Node.kt
@@ -2,6 +2,6 @@ package Models
 
 data class Node (
     val number: Int,
-    val x : Int,
-    val y : Int
+    val x : Double,
+    val y : Double
 )
